# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: workflow web API
  description: >-
    The read-and-configure API behind `workflow --web`. It exposes the same
    information the terminal UI shows — the assigned issues, the current branch,
    the working tree, the pull request and its CI, and where it announces — and
    lets the whole configuration file be read and written. It reuses the same domain seams the
    TUI and the `status`/`reviews` CLI commands use; it does not write anything
    but the configuration file, and stores nothing between requests.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:7000
    description: The local server, bound to the loopback interface only.
...
paths:
...
  /api/events:
    get:
      operationId: streamEvents
      summary: A stream of state snapshots, pushed on connect and as they change.
      description: >-
        A Server-Sent Events stream (text/event-stream). Each message is a
        "snapshot" event whose data is a Snapshot: the current issues, branch,
        changes, review, and messaging destination. The browser subscribes once and
        updates from the pushes rather than polling.
      tags: [events]
      parameters:
        - name: view
          in: query
          description: The configured view's name whose issues to include; the default view when omitted.
          required: false
          schema: {type: string}
      responses:
        "200":
          description: The event stream; each message's data is a Snapshot.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/Snapshot"
        default: {$ref: "#/components/responses/Problem"}
components:
  responses:
    Problem:
      description: An RFC 9457 problem details object describing the failure.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: >-
        An RFC 9457 problem details object. The detail is safe to show and never
        carries a secret; code is a stable, machine-readable reason.
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
          description: A URI that identifies the problem type.
          example: https://jacob-delgado.github.io/workflow/docs/errors/#not-found
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        code:
          type: string
          description: A stable, machine-readable reason.
          enum: [bad_request, not_found, conflict, unprocessable, unreachable, internal]
      additionalProperties: false
...
    Snapshot:
      type: object
      description: >-
        The full read state carried by one event-stream message: everything the
        cockpit shows, together.
      required: [issues, branch, changes, review, messaging, branches]
      properties:
        issues: {$ref: "#/components/schemas/IssuesPage"}
        branch: {$ref: "#/components/schemas/Branch"}
        changes: {$ref: "#/components/schemas/ChangeList"}
        review: {$ref: "#/components/schemas/Review"}
        messaging: {$ref: "#/components/schemas/MessagingDestination"}
        branches:
          type: array
          description: >-
            The local branches named for an issue — the record of what is in
            flight. The detail panels (branch, changes, review) describe the
            checked-out branch alone; this lists every issue that has a branch,
            so the issues list can mark them all in flight, not the one on HEAD.
          items: {$ref: "#/components/schemas/TaskBranch"}
      additionalProperties: false
    TaskBranch:
      type: object
      description: A local branch named for an issue, by the branch-name convention.
      required: [name, issue_key, current]
      properties:
        name:
          type: string
          description: The local branch name.
          example: fix/PROJ-412-redact-tokens
        issue_key:
          type: string
          description: >-
            The issue the branch is named for — a Jira key or a forge issue
            number, per the branch-name convention.
          example: PROJ-412
        current:
          type: boolean
          description: Whether this is the checked-out branch.
      additionalProperties: false
    IssuesPage:
      type: object
      required: [issues, total, start_at]
      properties:
        issues:
          type: array
          items: {$ref: "#/components/schemas/Issue"}
        total:
          type: integer
          description: How many issues the view matches in all.
        start_at:
          type: integer
          description: The index of the first issue in this page.
      additionalProperties: false
    Issue:
      type: object
      required: [key, summary, status, status_category, type]
      properties:
        key: {type: string, example: PROJ-412}
        summary: {type: string}
        status:
          type: string
          description: The human status name, such as "In Progress".
        status_category:
          $ref: "#/components/schemas/StatusCategory"
        type:
          type: string
          description: The issue type name, such as Bug or Story.
        priority:
          type: string
          description: May be empty.
      additionalProperties: false
    StatusCategory:
      type: string
      enum: [new, indeterminate, done]
...
    Branch:
      type: object
      required: [name, detached, head, upstream, ahead, behind, base, commits]
      properties:
        name: {type: string}
        detached:
          type: boolean
          description: Whether HEAD is detached rather than on a branch.
        head:
          type: string
          description: The commit SHA; empty before the first commit.
        upstream:
          type: string
          description: The tracking branch; empty if never pushed.
        ahead: {type: integer}
        behind: {type: integer}
        base:
          type: string
          description: The base branch, such as origin/main.
        commits:
          type: array
          description: The commits on this branch since its base.
          items: {$ref: "#/components/schemas/Commit"}
      additionalProperties: false
    Commit:
      type: object
      required: [hash, subject]
      properties:
        hash: {type: string}
        subject: {type: string}
      additionalProperties: false
    ChangeList:
      type: object
      required: [changes]
      properties:
        changes:
          type: array
          items: {$ref: "#/components/schemas/Change"}
      additionalProperties: false
    Change:
      type: object
      required: [path, kind, staged, has_unstaged, conflicted]
      properties:
        path: {type: string}
        original_path:
          type: string
          description: The former path for a rename or copy; empty otherwise.
        kind:
          type: string
          description: What kind of change this is.
          enum: [modified, new, deleted, renamed, copied, untracked, conflicted]
        staged:
          type: boolean
          description: Whether the change (or part of it) is staged.
        has_unstaged:
          type: boolean
          description: Whether there are unstaged edits to the same path.
        conflicted:
          type: boolean
      additionalProperties: false
    Review:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
          description: Whether a pull request was found for the branch.
        pull:
          nullable: true
          allOf: [{$ref: "#/components/schemas/PullRequest"}]
        ci:
          nullable: true
          allOf: [{$ref: "#/components/schemas/CI"}]
      additionalProperties: false
...
    PullRequest:
      type: object
      required: [number, url, title, draft, approvals, changes_requested, mergeable]
      properties:
        number:
          type: integer
          description: The GitHub number or GitLab IID.
        url: {type: string}
        title: {type: string}
        draft: {type: boolean}
        approvals: {type: integer}
        changes_requested: {type: boolean}
        mergeable:
          type: string
          enum: [unknown, clean, conflicts]
      additionalProperties: false
    CI:
      type: object
      required: [state, total, done, failed, checks]
      properties:
        state: {$ref: "#/components/schemas/CIState"}
        total:
          type: integer
          description: Total checks; 0 on a forge that reports the pipeline whole.
        done: {type: integer}
        failed: {type: integer}
        checks:
          type: array
          items: {$ref: "#/components/schemas/Check"}
      additionalProperties: false
    Check:
      type: object
      required: [name, state, url]
      properties:
        name: {type: string}
        state: {$ref: "#/components/schemas/CIState"}
        url: {type: string}
      additionalProperties: false
    CIState:
      type: string
      enum: [none, running, passed, failed]
    MessagingDestination:
      type: object
      required: [service, configured, channel, channels, author]
      properties:
        service:
          type: string
          description: The service posts go to — Slack, Teams, Discord or webhook.
        configured:
          type: boolean
          description: >-
            Whether a transport is set up to post — a bot token or a webhook. A
            webhook service has no channel, so channel presence cannot stand in
            for this.
        channel:
          type: string
          description: The default channel; empty when a webhook carries its own.
        channels:
          type: array
          description: The default first, then the alternates.
          items: {type: string}
        author:
          type: string
          description: Who a post would come from; empty if not resolved.
      additionalProperties: false
...